refactor : 채팅삭제 리팩토링

+++ src/main/java/com/example/springboot/dao/ChatroomDAOImpl.java
@@ -42,16 +42,19 @@ public class ChatroomDAOImpl implements ChatroomDAO {
     /**
      * 채팅방 삭제
      * <p>
-     * 채팅방, 채팅내역을 삭제하는 메소드이다.
+     * 채팅방을 삭제하는 메소드이다.
+     *
      * @param id
      * @param chatroomId
      */
     @Override
-    public void exitChatroom(String id, Long chatroomId) {
-        chatroomRepository.deleteById(chatroomId);
-        chatroomGroupRepository.deleteByChatroomId(chatroomId);
-        messageRepository.deleteByChatroomId(chatroomId);
+    public void deleteById(String id, Long chatroomId) {
+        if (id == null || id.isEmpty()) {
+            throw new ErrorRequiredValueValidationException(new StringBuilder("id는 "), ExceptionCodeEnum.NONEXISTENT_REQUIRED_VALUE);
+        }
 
-        log.info("exitChatroom chatroomId : {}", chatroomId);
+        log.info("exitChatroom id : {}, chatroomId : {}", id, chatroomId);
+
+        chatroomRepository.deleteById(chatroomId);
     }
 }
